Fix fast refresh compatibility for unit
